Fix google analytics
