+++ index.d.ts
@@ -1,0 +1,3 @@
+/// <reference types="gtag.js" />
+
+declare module "gtag.js";
