+++ package.json
@@ -19,6 +19,7 @@
     "@mantine/next": "^5.5.0",
     "@mantine/notifications": "^5.5.4",
     "@reduxjs/toolkit": "^1.8.5",
+    "@types/gtag.js": "^0.0.12",
     "axios": "^0.27.2",
     "axios-auth-refresh": "^3.3.3",
     "global": "^4.4.0",
