+++ pages/_document.tsx
@@ -1,5 +1,6 @@
 import { createGetInitialProps } from "@mantine/next";
 import Document, { Head, Html, Main, NextScript } from "next/document";
+import Script from "next/script";
 
 const getInitialProps = createGetInitialProps();
 
@@ -11,6 +12,25 @@ export default class _Document extends Document {
       <Html>
         <Head>
           <meta name="theme-color" content="#2c2f33" />
+          {/* Global Site Tag (gtag.js) - Google Analytics */}
+          <Script
+            strategy="afterInteractive"
+            src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GA_MEASUREMENT_ID}`}
+          />
+          <Script
+            id="gtag-init"
+            strategy="afterInteractive"
+            dangerouslySetInnerHTML={{
+              __html: `
+                window.dataLayer = window.dataLayer || [];
+                function gtag(){dataLayer.push(arguments);}
+                gtag('js', new Date());
+                gtag('config', '${process.env.NEXT_PUBLIC_GA_MEASUREMENT_ID}', {
+                  page_path: window.location.pathname,
+                });
+              `,
+            }}
+          />
           <link rel="manifest" href="/manifest.json" />
           {/* <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Quicksand&display=optional"/> */}
           <link
